ref: move Navigation and working directory to navigation.js
